Add leerling select fragment Database implementation of Student Fetching Student en persisting from backend (waiting for working endpoint) Basic layouting

+++ app/src/main/java/be/hogent/tile3/rubricapplication/model/Docent.kt
@@ -21,5 +21,5 @@ data class Docent(
     val naam: String,
 
     @ColumnInfo(name = "opleidingsOnderdeelId")
-    val opleidingsOnderdeelId: Int
+    val opleidingsOnderdeelId: Long
 )
